move upgrade-guest-form to settings dir
