+++ app/features/settings/components/upgrade-guest-form.tsx
@@ -11,7 +11,7 @@ import { Input } from '~/components/ui/input';
 import { Label } from '~/components/ui/label';
 import { useToast } from '~/hooks/use-toast';
 import { buildEmailValidator } from '~/lib/validation';
-import { useUserStore } from '../user/user-provider';
+import { useUserStore } from '../../user/user-provider';
 
 type FormValues = { email: string; password: string; confirmPassword: string };
